chore(function-to-string): use babel instead of recast
babel is already a dependency widely used in whole autodocs. Dropping
recast to hopefully reduce runtime bundle
