+++ packages/wix-storybook-utils/package.json
@@ -50,6 +50,10 @@
   },
   "dependencies": {
     "@babel/core": "^7.4.3",
+    "@babel/traverse": "^7.4.3",
+    "@babel/types": "^7.4.3",
+    "@babel/parser": "^7.4.3",
+    "@babel/generator": "^7.4.3",
     "@babel/plugin-proposal-class-properties": "^7.4.0",
     "@babel/plugin-syntax-jsx": "^7.2.0",
     "babel-runtime": "^6.25.0",
@@ -70,7 +74,6 @@
     "react-motion": "^0.5.2",
     "react-remarkable": "^1.1.3",
     "react-syntax-highlighter": "^9.0.0",
-    "recast": "^0.14.7",
     "wix-ui-icons-common": "^2.0.24"
   },
   "devDependencies": {
