+++ packages/wix-storybook-utils/src/ComponentSource/function-to-string.ts
@@ -1,0 +1,51 @@
+import * as types from '@babel/types';
+import * as parser from '@babel/parser';
+import generator from '@babel/generator';
+import traverse from '@babel/traverse';
+
+const ensureShorthandProperties = ({ ast, parentPath, scope }) =>
+  traverse(
+    ast,
+    {
+      Property(path) {
+        const { key, value } = path.node;
+        if (key.test === value.test) {
+          path.node.shorthand = true;
+        }
+      },
+    },
+    scope,
+    parentPath,
+  );
+
+const functionToString = prop => {
+  if (typeof prop !== 'function') {
+    return prop;
+  }
+
+  const ast = parser.parseExpression(prop.toString());
+
+  if (types.isArrowFunctionExpression(ast)) {
+    return prop;
+  }
+
+  const arrowFunctionBody =
+    ast.body.body.length === 1 && types.isReturnStatement(ast.body.body[0])
+      ? ast.body.body[0].argument
+      : ast.body;
+
+  ensureShorthandProperties({
+    ast: arrowFunctionBody,
+    parentPath: ast,
+    scope: ast,
+  });
+
+  const arrowFuncExpr = types.arrowFunctionExpression(
+    ast.params,
+    arrowFunctionBody,
+  );
+
+  return generator(arrowFuncExpr, { tabWidth: 2 }).code;
+};
+
+export default functionToString;
